@ Runtime bring-up + render-pipeline matching (menu boots+renders)
Runtime fixes:
- MarDirectorSetupObjects: unk40 push uses drawInit (TFrmGXSet, filter 0x8)
  not drawBufferGroup -> TFrmGXSet::perform now populates the per-frame
  TGraphics mRenderMode before the EFB present copy (was GXGetYScaleFactor
  hang / black menu). Verified vs original asm.
- Application gameLoop: fix C_MTXOrtho bottom/right arg swap (efbHeight vs
  fbWidth) -> screen fade now covers full screen, not just top ~2/3.
- (+ earlier bring-up fixes: scene-tree list type, makeObjAppeared
  recursion, MarioDraw/Cap/WaterGun/Yoshi init, PerformLists.bin buffer,
  wire back-pointer, camera unk0/gpCubeCamera, perform-list dispatch.)

Matching:
- JDREfbCtrl: 6/8 byte-identical (setTexAttb batched-store removing the
  char[0x10] hack; TEfbCtrlDisp::perform .get()+arg locals). 2 residuals
  = dead-param-reg / stack-padding artifacts.
- ScrnFader: 15/16 byte-identical (draw_wipe_box, load); update 99.94%.
- (+ Enemy/NPC/Player/MapObj/MSound reconstructions.)

Docs:
- docs/AUDIT.md: new strict manual-equivalence audit log (supersedes the
  loose .agent_verdicts stamps) + SYS-1 (dtk-split COMDAT artifact).
- docs/MWCC.md: dead-parameter register-reservation lever (2 TUs).
@

+++ include/Player/MarioMain.hpp
@@ -556,7 +556,7 @@ public:
 		TParamRT<s16> mRotGravity;
 		TParamRT<f32> mRotBrake;
 		TParamRT<f32> mJumpRate;
-		TParamRT<s32> mSwingRate;
+		TParamRT<f32> mSwingRate;
 		TParamRT<f32> mWireJumpAccelControl;
 		TParamRT<f32> mWireJumpSlideControl;
 		TParamRT<f32> mWireJumpMult;
